refactor: refactor by copilot comments
